% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{att.info}
\alias{att.info}
\title{the attribute information of the gff file}
\format{An object of class \code{NULL} of length 0.}
\usage{
att.info
}
\value{
the attribute information of the gff file
}
\description{
the attribute information of the gff file
}
\examples{
att.info=c()
}
\keyword{datasets}
