% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{chosen.column.indicies}
\alias{chosen.column.indicies}
\title{the chosen column of gff file (to remove unused columns from analysis)}
\format{An object of class \code{NULL} of length 0.}
\usage{
chosen.column.indicies
}
\value{
the chosen column of gff file (to remove unused columns from analysis)
}
\description{
the chosen column of gff file (to remove unused columns from analysis)
}
\examples{
chosen.column.indicies=c()
}
\keyword{datasets}
