% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{compute.database.exon.details}
\alias{compute.database.exon.details}
\title{look for the genomic location of the indel sequence in the annotated gff gene models}
\format{An object of class \code{logical} of length 1.}
\usage{
compute.database.exon.details
}
\value{
look for the genomic location of the indel sequence in the annotated gff gene models
}
\description{
look for the genomic location of the indel sequence in the annotated gff gene models
}
\examples{
compute.database.exon.details=TRUE
}
\keyword{datasets}
