% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{coverage.file}
\alias{coverage.file}
\title{coverage file to calculate sequence coverage for indel regions with BAM files}
\format{An object of class \code{character} of length 1.}
\usage{
coverage.file
}
\value{
coverage file to calculate sequence coverage for indel regions with BAM files
}
\description{
coverage file to calculate sequence coverage for indel regions with BAM files
}
\examples{
coverage.file= "alignment.bam"
}
\keyword{datasets}
