% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{dataDir}
\alias{dataDir}
\title{the directly where data is stored}
\format{An object of class \code{character} of length 1.}
\usage{
dataDir
}
\value{
the directly where data is stored
}
\description{
the directly where data is stored
}
\examples{
dataDir=""
}
\keyword{datasets}
