% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\name{find.indels.location.by.location}
\alias{find.indels.location.by.location}
\title{find the place (exon) of the indel sequence in annotation and predicted genes}
\usage{
find.indels.location.by.location(chosen.gene, chose.pacbio.transcript.id,
  indel.seq, pacbio.protein.sequences, gene.df.pacbio,
  reference.gff.info.chosen, pacbio.isof.ids,
  compute.query.exon.details = FALSE, compute.database.exon.details = FALSE)
}
\arguments{
\item{chosen.gene}{the gene of interest}

\item{chose.pacbio.transcript.id}{chosen gene's prediction name}

\item{indel.seq}{the chosen gene's alternative sequences}

\item{pacbio.protein.sequences}{the protein sequence of the predicted gene}

\item{gene.df.pacbio}{the data frame of the predicted genes (Gff)}

\item{reference.gff.info.chosen}{the data frame of the annnotated genes}

\item{pacbio.isof.ids}{the predicted gene's prediction id}

\item{compute.query.exon.details}{whether to find the neoantigen's position in predictions}

\item{compute.database.exon.details}{whether to find the neoantigen's position in annotations}
}
\value{
a vector of the predicted gene's genomic locations and locations in annotations
}
\description{
find the place (exon) of the indel sequence in annotation and predicted genes
}
