% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{gene.df.pacbio}
\alias{gene.df.pacbio}
\title{the data frame of predicted gene models}
\format{An object of class \code{data.frame} with 0 rows and 0 columns.}
\usage{
gene.df.pacbio
}
\value{
the data frame of predicted gene models
}
\description{
the data frame of predicted gene models
}
\examples{
gene.df.pacbio=data.frame()
}
\keyword{datasets}
