% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{gff.file}
\alias{gff.file}
\title{the gff file in the dat frame mode}
\format{An object of class \code{data.frame} with 0 rows and 0 columns.}
\usage{
gff.file
}
\value{
the gff file in the dat frame mode
}
\description{
the gff file in the dat frame mode
}
\examples{
gff.file=data.frame()
}
\keyword{datasets}
