% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{gff.names}
\alias{gff.names}
\title{the gff file's gene names}
\format{An object of class \code{character} of length 1.}
\usage{
gff.names
}
\value{
the gff file's gene names
}
\description{
the gff file's gene names
}
\examples{
gff.names="prediction.gff"
}
\keyword{datasets}
