% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\name{gffRead}
\alias{gffRead}
\title{read the gff gene prediction files}
\usage{
gffRead(gffFile, nrows = -1)
}
\arguments{
\item{gffFile}{the name of input file}

\item{nrows}{the num of rows to skip}
}
\value{
the data frame of gff input file
}
\description{
read the gff gene prediction files
}
