% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\name{is.installed}
\alias{is.installed}
\title{check if a package is installed or not}
\usage{
is.installed(mypkg)
}
\arguments{
\item{mypkg}{the name of a R package}
}
\value{
whether the package is installed, if not, then install it
}
\description{
check if a package is installed or not
}
\examples{
is.installed('seqinr')
}
