% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{mainDir}
\alias{mainDir}
\title{the directly where code and example data are stored}
\format{An object of class \code{character} of length 1.}
\usage{
mainDir
}
\value{
the directly where code and example data are stored
}
\description{
the directly where code and example data are stored
}
\examples{
mainDir="data"
}
\keyword{datasets}
