% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{max.mismatch.size}
\alias{max.mismatch.size}
\title{at most 1000 amino is different from the annotation}
\format{An object of class \code{numeric} of length 1.}
\usage{
max.mismatch.size
}
\value{
at most 1000 amino is different from the annotation
}
\description{
at most 1000 amino is different from the annotation
}
\examples{
max.mismatch.size=1000
}
\keyword{datasets}
