% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{min.mismatch.size}
\alias{min.mismatch.size}
\title{at least 3 amino acids must be different from the annotation}
\format{An object of class \code{numeric} of length 1.}
\usage{
min.mismatch.size
}
\value{
at least 3 amino acids must be different from the annotation
}
\description{
at least 3 amino acids must be different from the annotation
}
\examples{
min.mismatch.size=3
}
\keyword{datasets}
