% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{minimum.flanking.region.size}
\alias{minimum.flanking.region.size}
\title{mimimum anchor size in the indel region}
\format{An object of class \code{numeric} of length 1.}
\usage{
minimum.flanking.region.size
}
\value{
mimimum anchor size in the indel region
}
\description{
mimimum anchor size in the indel region
}
\examples{
minimum.flanking.region.size = 10
}
\keyword{datasets}
