% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{minimum.sequence.similarity}
\alias{minimum.sequence.similarity}
\title{the alternative isoform and reference protein sequence must be 50 percentage or more overlap}
\format{An object of class \code{numeric} of length 1.}
\usage{
minimum.sequence.similarity
}
\value{
the alternative isoform and reference protein sequence must be 50 percentage or more overlap
}
\description{
the alternative isoform and reference protein sequence must be 50 percentage or more overlap
}
\examples{
minimum.sequence.similarity     = 0.5
}
\keyword{datasets}
