% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\name{neoantigenR.get.peptides}
\alias{neoantigenR.get.peptides}
\title{compute neoantigens using the novel algorithm after the preparation steps}
\usage{
neoantigenR.get.peptides()
}
\value{
no values will be returned
}
\description{
compute neoantigens using the novel algorithm after the preparation steps
}
