% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\name{neoantigenR.initialize}
\alias{neoantigenR.initialize}
\title{initialize the parameters and functions for neoantigen prediction}
\usage{
neoantigenR.initialize()
}
\value{
global variables will be set after initialization
}
\description{
initialize the parameters and functions for neoantigen prediction
}
