% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{org}
\alias{org}
\title{the organism used to search protein database}
\format{An object of class \code{character} of length 1.}
\usage{
org
}
\value{
the organism used to search protein database
}
\description{
the organism used to search protein database
}
\examples{
org="hg19"
}
\keyword{datasets}
