% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{output.folder}
\alias{output.folder}
\title{the folder of the output files}
\format{An object of class \code{character} of length 1.}
\usage{
output.folder
}
\value{
the folder of the output files
}
\description{
the folder of the output files
}
\examples{
output.folder = "analysis"
}
\keyword{datasets}
