% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{pacbio.gencode.overlapping.file}
\alias{pacbio.gencode.overlapping.file}
\title{the predicted gff file intersecting with reference gene annotation file by BEDTools}
\format{An object of class \code{character} of length 1.}
\usage{
pacbio.gencode.overlapping.file
}
\value{
the predicted gff file intersecting with reference gene annotation file by BEDTools
}
\description{
the predicted gff file intersecting with reference gene annotation file by BEDTools
}
\examples{
pacbio.gencode.overlapping.file	 = "bedtool.intersect.overlaps.txt"
}
\keyword{datasets}
