% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{pacbio.gff}
\alias{pacbio.gff}
\title{the predicted gff file}
\format{An object of class \code{character} of length 1.}
\usage{
pacbio.gff
}
\value{
the predicted gff file
}
\description{
the predicted gff file
}
\examples{
pacbio.gff="cufflinks.gff"
}
\keyword{datasets}
