% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{pacbio.isof.ids}
\alias{pacbio.isof.ids}
\title{the isoform or gene names from predicted gff model}
\format{An object of class \code{NULL} of length 0.}
\usage{
pacbio.isof.ids
}
\value{
the isoform or gene names from predicted gff model
}
\description{
the isoform or gene names from predicted gff model
}
\examples{
pacbio.isof.ids=c()
}
\keyword{datasets}
