% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{percOverlap}
\alias{percOverlap}
\title{the alternative isoform and reference protein sequence must be 80 percentage or more overlap}
\format{An object of class \code{numeric} of length 1.}
\usage{
percOverlap
}
\value{
the alternative isoform and reference protein sequence must be 80 percentage or more overlap
}
\description{
the alternative isoform and reference protein sequence must be 80 percentage or more overlap
}
\examples{
percOverlap	= 0.8
}
\keyword{datasets}
