% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{protein.database.file.name}
\alias{protein.database.file.name}
\title{the path for reference protein database file (Swiss-uniprot)}
\format{An object of class \code{character} of length 1.}
\usage{
protein.database.file.name
}
\value{
the path for reference protein database file (Swiss-uniprot)
}
\description{
the path for reference protein database file (Swiss-uniprot)
}
\examples{
protein.database.file.name="swissuniprots.fasta"
}
\keyword{datasets}
