% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{reference.gff.file}
\alias{reference.gff.file}
\title{the path for reference annotated gene database file}
\format{An object of class \code{character} of length 1.}
\usage{
reference.gff.file
}
\value{
the path for reference annotated gene database file
}
\description{
the path for reference annotated gene database file
}
\examples{
reference.gff.file="gencode.v19.annotation.gff3"
}
\keyword{datasets}
