% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{reference.gff.info.chosen}
\alias{reference.gff.info.chosen}
\title{the annotated gene models for chosen genes}
\format{An object of class \code{data.frame} with 0 rows and 0 columns.}
\usage{
reference.gff.info.chosen
}
\value{
the annotated gene models for chosen genes
}
\description{
the annotated gene models for chosen genes
}
\examples{
reference.gff.info.chosen=data.frame()
}
\keyword{datasets}
