% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{reference.gff.info}
\alias{reference.gff.info}
\title{the annotated gene in data frame mode with attributes splitted}
\format{An object of class \code{data.frame} with 0 rows and 0 columns.}
\usage{
reference.gff.info
}
\value{
the annotated gene in data frame mode with attributes splitted
}
\description{
the annotated gene in data frame mode with attributes splitted
}
\examples{
reference.gff.info=data.frame()
}
\keyword{datasets}
