% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{reference.gff}
\alias{reference.gff}
\title{the annotated gene in data frame mode}
\format{An object of class \code{data.frame} with 0 rows and 0 columns.}
\usage{
reference.gff
}
\value{
the annotated gene in data frame mode
}
\description{
the annotated gene in data frame mode
}
\examples{
reference.gff=data.frame()
}
\keyword{datasets}
