% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{refseq.pacbio.name.ids}
\alias{refseq.pacbio.name.ids}
\title{the predicted gene model's gene ids (prediction names)}
\format{An object of class \code{NULL} of length 0.}
\usage{
refseq.pacbio.name.ids
}
\value{
the predicted gene model's gene ids (prediction names)
}
\description{
the predicted gene model's gene ids (prediction names)
}
\examples{
refseq.pacbio.name.ids=c()
}
\keyword{datasets}
