% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{sample.pacbio.gencode.id.mapping}
\alias{sample.pacbio.gencode.id.mapping}
\title{the mapping between predicted gene names and reference annotated gene names}
\format{An object of class \code{data.frame} with 0 rows and 0 columns.}
\usage{
sample.pacbio.gencode.id.mapping
}
\value{
the mapping between predicted gene names and reference annotated gene names
}
\description{
the mapping between predicted gene names and reference annotated gene names
}
\examples{
sample.pacbio.gencode.id.mapping=data.frame()
}
\keyword{datasets}
