% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{target.gene.list}
\alias{target.gene.list}
\title{the gene list finally chosen for neoantigen analysis}
\format{An object of class \code{NULL} of length 0.}
\usage{
target.gene.list
}
\value{
the gene list finally chosen for neoantigen analysis
}
\description{
the gene list finally chosen for neoantigen analysis
}
\examples{
target.gene.list=c("MET")
}
\keyword{datasets}
