% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{uniprot.human.database}
\alias{uniprot.human.database}
\title{the uniprot human dataset in data frame}
\format{An object of class \code{data.frame} with 0 rows and 0 columns.}
\usage{
uniprot.human.database
}
\value{
the uniprot human dataset in data frame
}
\description{
the uniprot human dataset in data frame
}
\examples{
uniprot.human.database=data.frame()
}
\keyword{datasets}
