% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{uniprot.protein.names.new}
\alias{uniprot.protein.names.new}
\title{the human protein names from uniprot}
\format{An object of class \code{character} of length 1.}
\usage{
uniprot.protein.names.new
}
\value{
the human protein names from uniprot
}
\description{
the human protein names from uniprot
}
\examples{
uniprot.protein.names.new=""
}
\keyword{datasets}
