% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{uniprot.protein.swissprot.index}
\alias{uniprot.protein.swissprot.index}
\title{the human protein names from uniprot swissport (only the rows or indexes out of the entire uniprot file)}
\format{An object of class \code{character} of length 1.}
\usage{
uniprot.protein.swissprot.index
}
\value{
the human protein names from uniprot swissport (only the rows or indexes out of the entire uniprot file)
}
\description{
the human protein names from uniprot swissport (only the rows or indexes out of the entire uniprot file)
}
\examples{
uniprot.protein.swissprot.index=""
}
\keyword{datasets}
