% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{use.reference.annotation.gff}
\alias{use.reference.annotation.gff}
\title{indicate whether we will use reference gene annotation or now}
\format{An object of class \code{logical} of length 1.}
\usage{
use.reference.annotation.gff
}
\value{
indicate whether we will use reference gene annotation or now
}
\description{
indicate whether we will use reference gene annotation or now
}
\examples{
use.reference.annotation.gff=TRUE
}
\keyword{datasets}
