% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{validate.in.database.boolean}
\alias{validate.in.database.boolean}
\title{whether to validate indels in the annotated protein database}
\format{An object of class \code{logical} of length 1.}
\usage{
validate.in.database.boolean
}
\value{
whether to validate indels in the annotated protein database
}
\description{
whether to validate indels in the annotated protein database
}
\examples{
validate.in.database.boolean=TRUE
}
\keyword{datasets}
