% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\docType{data}
\name{write.dna.seq.by.reference.genome}
\alias{write.dna.seq.by.reference.genome}
\title{indicate whether we will write dna sequence for protein translation}
\format{An object of class \code{logical} of length 1.}
\usage{
write.dna.seq.by.reference.genome
}
\value{
indicate whether we will write dna sequence for protein translation
}
\description{
indicate whether we will write dna sequence for protein translation
}
\examples{
write.dna.seq.by.reference.genome	=	TRUE
}
\keyword{datasets}
