% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenR.R
\name{write.dna.sequence.reference}
\alias{write.dna.sequence.reference}
\title{write the dna sequence of a predicted gene based on reference genome object}
\usage{
write.dna.sequence.reference(Hsapiens, gene.df, file.name, gene.name)
}
\arguments{
\item{Hsapiens}{the name of input file}

\item{gene.df}{a data frame (gff) of predicted gene's coordinates}

\item{file.name}{the output file name to be written}

\item{gene.name}{the gene of interest for dna sequence extraction}
}
\value{
a dna sequence file will be written to output file
}
\description{
write the dna sequence of a predicted gene based on reference genome object
}
